Move blocklist code into tests/common/blocklist.py
